requirements.txt
pandas~=2.3.1
numpy~=2.3.2
faiss-cpu~=1.11.0.post1
pypdf~=5.9.0
sentence-transformers~=5.1.0